move var_bound_add to util.py
