+++ pyomo/contrib/mindtpy/initialization.py
@@ -15,6 +15,7 @@ from pyomo.solvers.plugins.solvers.persistent_solver import PersistentSolver
 from pyomo.contrib.mindtpy.nlp_solve import (solve_NLP_subproblem,
                                              handle_NLP_subproblem_optimal, handle_NLP_subproblem_infeasible,
                                              handle_NLP_subproblem_other_termination)
+from pyomo.contrib.mindtpy.util import var_bound_add
 
 
 def MindtPy_initialize_master(solve_data, config):
@@ -24,7 +25,6 @@ def MindtPy_initialize_master(solve_data, config):
     """
     # if single tree is activated, we need to add bounds for unbounded variables in nonlinear constraints to avoid unbounded master problem.
     if config.single_tree == True:
-        from pyomo.contrib.mindtpy.single_tree import var_bound_add
         var_bound_add(solve_data, config)
 
     m = solve_data.mip = solve_data.working_model.clone()
