+++ pyomo/contrib/mindtpy/single_tree.py
@@ -238,25 +238,3 @@ class LazyOACallback_cplex(LazyConstraintCallback):
             self.handle_lazy_NLP_subproblem_other_termination(fixed_nlp, fixed_nlp_result.solver.termination_condition,
                                                               solve_data, config)
 
-
-def var_bound_add(solve_data, config):
-    """This function will add bound for variables in nonlinear constraints if they are not bounded.
-       This is to avoid an unbound master problem in the LP/NLP algorithm.
-    """
-    m = solve_data.working_model
-    MindtPy = m.MindtPy_utils
-    for c in MindtPy.constraint_list:
-        if c.body.polynomial_degree() not in (1, 0):
-            for var in list(EXPR.identify_variables(c.body)):
-                if var.has_lb() and var.has_ub():
-                    continue
-                elif not var.has_lb():
-                    if var.is_integer():
-                        var.setlb(-config.integer_var_bound)
-                    else:
-                        var.setlb(-config.continuous_var_bound)
-                elif not var.has_ub():
-                    if var.is_integer():
-                        var.setub(config.integer_var_bound)
-                    else:
-                        var.setub(config.continuous_var_bound)
